# Closes #1901: Reduce SymEntry creation overheads